+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # CasualOS Changelog
 
+## V3.0.3
+
+#### Date: 3/22/2022
+
+### :rocket: Improvements
+
+-   Added the `os.getAverageFrameRate()` function.
+    -   This function is useful for calculating the number of times that the 3D views have updated in the last second.
+    -   Returns a promise that resolves with the current frame rate value.
+-   `AUX_PLAYER_MODE`: The player mode that this instance should indicate to scripts.
+    -   `"player"` indicates that the inst is supposed to be for playing AUXes while `"builder"` indicates that the inst is used for building AUXes.
+    -   Defaults to `"builder"`.
+    -   This value is exposed via the object returned from `os.version()`.
+        -   See the documentation on `os.version()` for more information.
+-   Added a button that offers to redirect to a static instance after a 25 second loading timeout.
+    -   The redirect will send the user to `static.{common_host}` so `casualos.com` will redirect to `static.casualos.com` and `stable.casualos.com` will redirect to `static.casualos.com`.
+
 ## V3.0.2
 
 #### Date: 3/16/2022